refactor: 템플릿 설정 컴포넌트 분리
